requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
alpaca-py>=0.13.0
pytz>=2024.0
requests>=2.0.0